% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwc.R
\name{fasta_wc}
\alias{fasta_wc}
\title{Cound words in fasta file}
\usage{
fasta_wc(k = 1, file = "", text = NULL)
}
\arguments{
\item{k}{Word size}

\item{file}{FASTA filename}

\item{text}{FASTA data as character vector}
}
\description{
Cound words in fasta file
}
\examples{
# Get system FASTA file holding the Agathis dammara chloroplast
# protein seqences
aa_file <- system.file("Agathis_dammara.chloroplast.faa", package=("zwc"))
# Count 1-mers (character frequency)
fasta_wc(aa_file, k=1)
# Count 2-mers
fasta_wc(aa_file, k=2)

# A raw string of text can be given
fasta_wc(k=2, text="AAATTTTTAAATATATATATATATATATATA")

# Or a FASTA file as a string
data(mito_dna)
fasta_wc(k=2, text=mito_dna) 
}
