% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fasta_wf}
\alias{fasta_wf}
\title{Count words in fasta file}
\usage{
fasta_wf(k, file)
}
\arguments{
\item{k}{Integer expressing word size}

\item{file}{Fasta filename}
}
\description{
Count words in fasta file
}

