% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwc.R
\name{hist.wordcount}
\alias{hist.wordcount}
\title{Histograph of counts}
\usage{
\method{hist}{wordcount}(x, ...)
}
\arguments{
\item{x}{a class wordcount entity}

\item{...}{other arguments for hist()}
}
\value{
a histogram plot
}
\description{
Histograph of counts
}
