% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_aa}
\alias{mito_aa}
\title{Protein sequence from Arabidopsis thaliana mitochondria}
\format{Character vector}
\source{
TAIR10 Arabidopsis thaliana annotation
}
\usage{
mito_aa
}
\description{
Protein sequence from Arabidopsis thaliana mitochondria
}
\keyword{datasets}
