% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mito_dna}
\alias{mito_dna}
\title{DNA sequence from Arabidopsis thaliana mitochondria}
\format{Character vector}
\source{
TAIR10 Arabidopsis thaliana annotation
}
\usage{
mito_dna
}
\description{
DNA sequence from Arabidopsis thaliana mitochondria
}
\keyword{datasets}
