% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwc.R
\name{plot.wordcount}
\alias{plot.wordcount}
\title{Bar plot of counts}
\usage{
\method{plot}{wordcount}(x, ...)
}
\arguments{
\item{x}{a class wordcount entity}

\item{...}{other arguments for ggplot2::geom_bar()}
}
\value{
a bar plot
}
\description{
Bar plot of counts
}
