% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wc_from_file}
\alias{wc_from_file}
\title{Count words in fasta file}
\usage{
wc_from_file(k, file)
}
\arguments{
\item{k}{Integer expressing word size}

\item{file}{Fasta filename}
}
\description{
Count words in fasta file
}

