% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wc_from_string}
\alias{wc_from_string}
\title{Count words in fasta character array}
\usage{
wc_from_string(k, fasta_string)
}
\arguments{
\item{k}{Integer expressing word size}

\item{buffer}{A character array holding all fasta sequence}
}
\description{
Count words in fasta character array
}

