\name{zwc-package}
\alias{zwc-package}
\alias{zwc}
\docType{package}
\title{
    Count words in a FASTA file
}
\description{

    I wrote this program because I needed a fast fasta word counter and the
    EMBOSS `wordcount` utility is extremely slow.

    zwc is on the order of 10X to 100X faster than EMBOSS program, depending
    on the parameters and input data.

}
\details{
    No details available yet, sorry.
}
\author{
Maintainer: Zebulun Arendsee <arendsee@iastate.edu>
}
\keyword{ FASTA }
