% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zwc.R
\name{zwc}
\alias{zwc}
\title{Cound words in fasta file}
\usage{
zwc(k = 1, file = "", text = NULL)
}
\arguments{
\item{k}{Word size}

\item{file}{FASTA filename}

\item{text}{FASTA data as character vector}
}
\description{
Cound words in fasta file
}

